winex11: Give focus to the expected window when unmapping focused window.
SetForegroundWindow calls aren't reflected to the X11 side, and we're
sometimes becoming inconsistent between the X focused window and the
Win32 foreground window.

When a window gets unmapped, the window manager will send WM_TAKE_FOCUS
to a different window, which may not be the foreground window from the
Win32 point of view. Requesting input focus to our foreground before
unmapping the focused window effectively resyncs the state.