dxgi: Directly signal the frame latency fence.
Instead of going through the command queue again. That's because at
that point we just submitted the frame for presentation to Vulkan,
so the best approximation we have for when the frame will be
presented is now, not after the currently outstanding work in the
command queue is processed. It's still a rather poor approximation,
but for something better we need VK_KHR_present_wait.