libs: Add unmodified ARM builtin sources from compiler-rt
When building for ARM targets, these helper functions are needed.
When building with a llvm-mingw provided compiler, the default
compiler-rt builtins from that distribution get linked in
implicitly, but if building with a plain Clang (in MSVC mode),
Wine needs to supply these routines (in the same way as Wine
supplies e.g. __chkstk).

These are taken from the compiler-rt release 8.0.1; this was the
last version which was dual licensed under the MIT and University
of Illinois licenses. (Newer versions are under Apache 2.0 with
LLVM Exceptions.) The files are taken from the
compiler-rt/lib/builtins directory.