compiler-rt: Tweak ARM assembly ifdefs
Change __MINGW32__ ifdefs to _WIN32; in compiler-rt builtins, these
functions are only needed for mingw targets, as the functions
are expected to be provided by MSVC libraries in MSVC builds.

When used in Wine, these files are needed specifically to fulfill
that role in MSVC mode builds, so generalize the ifdefs.